Use `overflow: auto` for pre/code tags
On index.html there is a code tag which shows an example script tag to
include hogan.js on your site. This code tag is not scrollable and the
majority of the content is hidden. Scrolling is not possible, which
makes it really hard to copy the URL.

This change sets `overflow` to `auto` on `code` and `pre` elements. The
layout stays the same, but now the content is scrollable and it's
possible to copy the script tag.