Add hulk support for node exports format
Allow export of templates in node format, also supporting browserify/webpack processing

If we are creating a aggregated out put the export will expose a template map
like it is already doing to the global variable, but in this case we use
`module.export` and `global.` for browser interoperability.

If we are crearing one file per template it adds the template directly in the `module.exports`.